% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_graphs.R
\name{stochastic_explorer}
\alias{stochastic_explorer}
\title{Stochastic plot}
\usage{
stochastic_explorer(data_dir = "//wpia-hn2.hpc.dide.ic.ac.uk/vimc_stochastics")
}
\arguments{
\item{data_dir}{The location of the standardised stochastic folder
hierarchy; this can be a local path, a fully-qualified network path on
windows, or a mount point on linux or Mac.}
}
\description{
Launch a Shiny app to allow interactive plotting of
standardised stochastic data, burden estimates,
impacts, comparisons between touchstones, and
comparisons between modelling groups.
}
