% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{stone_extract}
\alias{stone_extract}
\title{Extract touchstone-relevant data}
\usage{
stone_extract(path, con)
}
\arguments{
\item{path}{Path to the import project root used for finding any local data.}

\item{con}{The active DBI connection for extracting any data.}
}
\value{
A list of named data frames and/or named values representing the extracted data.
}
\description{
Extract touchstone-relevant data from sources. This should be called with
\code{stoner::stone_extract(path, con}, from within the extract code of a
montagu-import, providing the same arguments provided to the montagu-import
extract function.
}
\details{
See the vignette for information about the specifics of the extraction.
}
