% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{stone_load}
\alias{stone_load}
\title{Load touchstone data to database}
\usage{
stone_load(transformed_data, con)
}
\arguments{
\item{transformed_data}{The transformed data from the transform stage. In a
form which adheres to the DB schema.}

\item{con}{Active connection to the DB which the data will be loaded to.}
}
\description{
Load touchstone-relevant transformed data into the database. This should be
called from the load code of a montagu-import, which must therefore be a
custom load, not an automatic one. Call \code{stoner::stone_load(transformed_data, con)}
with the same arguments provided to the montagu-import load function.
}
