% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{stone_prune}
\alias{stone_prune}
\title{Prune obsolete datasets}
\usage{
stone_prune(
  con,
  modelling_group = NULL,
  disease = NULL,
  touchstone = NULL,
  scenario = NULL,
  dry_run = TRUE,
  return_df = FALSE
)
}
\arguments{
\item{con}{DBI connection to database containing burden estimate sets}

\item{modelling_group}{Only prune specific modelling group(s).}

\item{disease}{Only prune specific disease(s)}

\item{touchstone}{Only prune specific touchstone(s)}

\item{scenario}{Only prune specific scenario description(s).}

\item{dry_run}{If true, only report details on what rows would be dropped.}

\item{return_df}{Used for testing or debugging; if true the function returns
a data frame of the burden estimate sets being considered for deletion, or
NULL if there is no work to do. If
false, the function returns the number of rows proposed to be deleted, which
may be zero.}
}
\description{
To save space on production, we could delete burden estimate sets
which are not a modelling_group's current estimate set for a
particular touchstone and scenario
}
