% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_files.R
\name{stone_stochastic_central}
\alias{stone_stochastic_central}
\title{Produce central average from stochastic files.}
\usage{
stone_stochastic_central(
  base,
  touchstone,
  disease,
  group,
  scenario,
  avg_method = mean
)
}
\arguments{
\item{base}{The folder in which the standardised stochastic files are found.}

\item{touchstone}{The touchstone name (for the graph title)}

\item{disease}{The disease, used for building the filename and graph title.}

\item{group}{The modelling group, used in the filename and graph title.}

\item{scenario}{The scenario to plot.}

\item{avg_method}{The function to use for averaging the outcomes, probably
mean (the default) or median.}
}
\description{
Create a central stochastic file from a set of standardised
files created by stone_stochastic_standardise. This is for
a single scenario, and we'll expect to find a file for each
country containing all the runs. We'll produce one file with
the mean (or median) of the stochastics, and all countries
in one file.
}
