% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_cert.R
\name{stone_stochastic_cert_verify}
\alias{stone_stochastic_cert_verify}
\title{Validate stochastic parameter certificate}
\usage{
stone_stochastic_cert_verify(con, certfile, modelling_group, touchstone)
}
\arguments{
\item{con}{DBI connection to production. Used for verifying certificate
against expected properties}

\item{certfile}{Name of the certificate file accompanying the estimates}

\item{modelling_group}{The modelling group who provided the certificate}

\item{touchstone}{The touchstone the group reports for this certificate.}
}
\description{
Montagu generates a certificate when a group uploads their stochastic
parameter set. Stoner can check the certificate applies to the right
touchstone and modelling group.
}
