% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_files.R
\name{stone_stochastic_make_meta}
\alias{stone_stochastic_make_meta}
\title{Produce \code{meta.csv} summary of the structure and
content of a standardised stochastic data folder.}
\usage{
stone_stochastic_make_meta(path)
}
\arguments{
\item{path}{The root folder of the stochastic data.}
}
\description{
Create a \code{meta.csv} file in the root of the standardised
stochastics. The columns contain scalars of \code{touchstone},
\code{disease}, \code{group}, \code{scenario} - and for each row, a
semi-colon-separated lists for \code{countries} and \code{outcomes}.
This is useful for making the stochastic explorer faster
on startup (otherwise it has to sample all of the files
each time you run it) - and also it is a good general
record of all the stochastic data we have.
}
\details{
This does mean that we should re-create the meta data
each time we make changes to the standardised stochastic
data though.
}
