% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_upload.R
\name{stone_stochastic_upload}
\alias{stone_stochastic_upload}
\title{Upload stochastic data to annex}
\usage{
stone_stochastic_upload(
  file,
  con,
  annex,
  modelling_group,
  disease,
  touchstone,
  is_cohort,
  is_under5
)
}
\arguments{
\item{file}{A csv file generated by stone_stochastic_process}

\item{con}{DBI connection to production (for argument validation)}

\item{annex}{DBI connection to annex, to receive stochastic uploads.}

\item{modelling_group}{The modelling group id}

\item{disease}{The disease}

\item{touchstone}{The touchstone (including version) for these estimates}

\item{is_cohort}{Set this to TRUE if the csv file is cohort-oriented, or
FALSE for calendar-year oriented.}

\item{is_under5}{Set this to TRUE if the csv file only considers ages
0-4 inclusive; FALSE if all ages are included.}
}
\description{
Upload stochastic data to annex
}
