% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_extract.R
\name{stone_test_extract}
\alias{stone_test_extract}
\title{Test the extracted data.}
\usage{
stone_test_extract(extracted_data)
}
\arguments{
\item{extracted_data}{A list of \code{data.frame}s and/or values,}
}
\description{
Run all generic tests on the provided extracted data. This should be
called from \code{test-extract.R} in a montagu-imports import. It tests
that the csv metadata is syntactically valid and as expected, and
that querying related database tables has also worked. After
calling \code{stoner::stone_text_extract(extracted_data)}, you should
write tests that are specific to the metadata for the particular
import you are writing.
}
\details{
See the vignette for information on the stoner-specific tests.
}
