% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_transform.R
\name{test_transform}
\alias{test_transform}
\title{Test the transformed data.}
\usage{
test_transform(transformed_data)
}
\arguments{
\item{transformed_data}{A list of \code{data.frame}s produced by the transform
function. They will have names matching database tables in montagu, and
columns matching the columns of those tables.}
}
\description{
Run all generic tests on the provided transformed data. This should be
called from \code{test-extract.R} in a montagu-imports import. It tests
that the csv metadata is syntactically valid and as expected, and
that querying related database tables has also worked. After
calling \code{stoner::text_extract(extracted_data)}, you should
write tests that are specific to the metadata for the particular
import you are writing.
}
\details{
See the vignette for information on the stoner-specific tests.
}
