% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{Transform extracted_data into transformed_data}
\usage{
transform(extracted_data)
}
\arguments{
\item{extracted_data}{The extracted data that was generated by the previous extract call}
}
\value{
A list of data frames, one for each table, with names matching the
database table names, and containing rows to be added to that table.
}
\description{
Transform touchstone-relevant extracted data into tables ready for
database import. This should be called from the transform code of montagu-import,
calling \code{stoner::transform(extracted_data)} with the same argument provided
to the montagu-import transform function.
}
\details{
For specific information on what steps the transform function does.
see the vignette.
}
