% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packit.R
\name{fetch_packit}
\alias{fetch_packit}
\title{Fetch packit}
\usage{
fetch_packit(
  packet_id,
  filename,
  server = "https://montagu.vaccineimpact.org/packit/api/"
)
}
\arguments{
\item{filename}{The filename of the file within the packit.}

\item{server}{By default, the URL to the packit API on Montagu,
but this can be set to other packit API's if we want.}

\item{packit_id}{The id of the packit containing the artefact.}
}
\value{
The filename of the temporary file which has been downloaded.
}
\description{
Download a file from a packit, for example an artefact containing
central burden estimates, which we could then plot on top of
stochastics. But more generally, we can use this for fetching any
file from packet (ie, from the Montagu Reporting Portal).
}
