% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_graphs.R
\name{stochastic_explorer}
\alias{stochastic_explorer}
\title{Stochastic plot}
\usage{
stochastic_explorer()
}
\description{
Launch a Shiny app to allow interactive plotting of
standardised stochastic data, burden estimates,
impacts, comparisons between touchstones, and
comparisons between modelling groups.
}
