% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_cert.R
\name{stone_stochastic_cert_verify}
\alias{stone_stochastic_cert_verify}
\title{Validate stochastic parameter certificate}
\usage{
stone_stochastic_cert_verify(
  con,
  certfile,
  modelling_group,
  touchstone,
  disease
)
}
\arguments{
\item{con}{DBI connection to production. Used for verifying certificate
against expected properties}

\item{certfile}{Name of the certificate file to be verified.}

\item{modelling_group}{The expected modelling group. We expect that the
owner of this certificate logged in as this modelling_group, when
uploading their stochastic parameter set.}

\item{touchstone}{The expected touchstone. We expect that the owner of
the provided certificate uploaded a stochastic parameter set to this
touchstone in order to get it.}

\item{disease}{The expected disease. We expect that the certificate
provided was uploaded into Montagu with reference to this disease.}
}
\description{
Montagu generates a certificate when a group uploads their stochastic
parameter set. Stoner can check the certificate applies to the right
touchstone and modelling group.
}
