% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_graphs.R
\name{stone_stochastic_graph}
\alias{stone_stochastic_graph}
\title{Stochastic plot}
\usage{
stone_stochastic_graph(
  base,
  touchstone,
  disease,
  group,
  country,
  scenario,
  outcome,
  ages = NULL,
  by_cohort = FALSE,
  log = FALSE
)
}
\arguments{
\item{base}{The folder in which the standardised stochastic files are found.}

\item{touchstone}{The touchstone name (for the graph title)}

\item{disease}{The disease, used for building the filename and graph title.}

\item{group}{The modelling group, used in the filename and graph title.}

\item{country}{The country to plot.}

\item{scenario}{The scenario to plot.}

\item{ages}{A vector of one or more ages to be selected and aggregated, or
if left as NULL, then all ages are used and aggregated.}

\item{by_cohort}{If TRUE, then age is subtracted from year to convert it to
year of birth before aggregating.}

\item{log}{If TRUE, then use a logged y-axis.}
}
\description{
Draw a stochastic plot showing all the different runs, with the mean,
median, 5\% and 95\% quantiles shown.
}
