% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_upload.R
\name{stone_stochastic_upload}
\alias{stone_stochastic_upload}
\title{Upload stochastic data to annex}
\usage{
stone_stochastic_upload(
  file,
  con,
  annex,
  modelling_group,
  disease,
  touchstone,
  is_cohort,
  is_under5,
  allow_new_database = FALSE,
  testing = FALSE
)
}
\arguments{
\item{file}{A pq or csv file generated by stone_stochastic_process}

\item{con}{DBI connection to production (for argument validation)}

\item{annex}{DBI connection to annex, to receive stochastic uploads.}

\item{modelling_group}{The modelling group id}

\item{disease}{The disease}

\item{touchstone}{The touchstone (including version) for these estimates}

\item{is_cohort}{Set this to TRUE if the csv file is cohort-oriented, or
FALSE for calendar-year oriented.}

\item{is_under5}{Set this to TRUE if the csv file only considers ages
0-4 inclusive; FALSE if all ages are included.}

\item{allow_new_database}{Create the stochastic_file database if it
does not exist. Should only be needed first time on a new database.}

\item{testing}{For internal use only.}
}
\description{
Upload the reduced csvs produced by stoner_stochastic_process into
the annex database.
}
